+++ src/components/templates/icons/user/user.jsx
@@ -3,8 +3,6 @@ import ProfilePhoto from "../components/ProfilePhoto";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./style.css";
 import CreateAvatar from "../components/CreateAvatar";
-import Portal from "../components/Portal";
-// import MydotsProfile from "../components/MydotsProfile";
 
 const UserIcon = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -18,14 +16,11 @@ const UserIcon = () => {
   return (
     <React.Fragment>
       <ProfilePhoto getData={getData} imageSrc={imageSrc} />
-      {/* <MydotsProfile getData={getData} imageSrc={imageSrc} /> */}
       <div id="createAvatarDiv" />
       {isOpen && !imageSrc && (
-        <Portal portalDiv="createAvatarDiv">
-          <div className="createAvatarDiv_content m-auto">
-            <CreateAvatar getData={getData} />
-          </div>
-        </Portal>
+        <div className="createAvatarDiv_content m-auto">
+          <CreateAvatar getData={getData} />
+        </div>
       )}
       <br />
     </React.Fragment>
